feat: add simple-verify test
