+++ test/assets/simple_proof_expected.json
