+++ test/assets/simple_proof_redacted.json
