+++ test/full-integration-swapi.spec.ts
@@ -1,7 +1,5 @@
 import { prove, verify } from '../src';
-import simple_proof_redacted from './simple_proof_redacted.json';
-
-const assert = require('assert');
+import simple_proof_redacted from './assets/simple_proof_redacted.json';
 
 (async function () {
   try {
@@ -28,42 +26,6 @@ const assert = require('assert');
     console.log(result);
     // @ts-ignore
     document.getElementById('full-integration-swapi').textContent =
-      JSON.stringify(result);
-  } catch (err) {
-    console.log('caught error from wasm');
-    console.error(err);
-  }
-})();
-
-(async function verify_simple() {
-  try {
-    const pem = `-----BEGIN PUBLIC KEY-----
-MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEBv36FI4ZFszJa0DQFJ3wWCXvVLFr
-cRzMG5kaTeHGoSzDu6cFqx3uEWYpFGo6C0EOUgf+mEgbktLrXocv5yHzKg==
------END PUBLIC KEY-----`;
-
-    const proof = {
-      notaryUrl: 'http://localhost',
-      ...simple_proof_redacted,
-    };
-
-    console.log(proof);
-
-    console.time('verify');
-    const result = await verify(proof, pem);
-    console.timeEnd('verify');
-
-    assert(result.serverName === 'example.com');
-    assert(
-      result.sent.includes(
-        'user-agent: XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX',
-      ),
-    );
-    assert(result.sent.includes('<h1>XXXXXXXXXXXXXX</h1'));
-
-    console.log(result);
-    // @ts-ignore
-    document.getElementById('verify_simple').textContent =
       JSON.stringify(result);
   } catch (err) {
     console.log('caught error from wasm');
