+++ test/test.ejs
@@ -12,8 +12,8 @@
     <script>
         global = globalThis //<- this should be enough
     </script>
-<div id="full-integration-swapi"></div>
-<div id="verify_simple"></div>
+    <div id="full-integration-swapi"></div>
+    <div id="simple-verify"></div>
 </body>
 
 </html>
