+++ webpack.web.dev.config.js
@@ -37,7 +37,8 @@ module.exports = [
     target: 'web',
     mode: isProd ? 'production' : 'development',
     entry: {
-      test: path.join(__dirname, 'test', 'full-integration-swapi.spec.ts'),
+      'full-integration-swapi.spec': path.join(__dirname, 'test', 'full-integration-swapi.spec.ts'),
+      'simple-verify': path.join(__dirname, 'test', 'simple-verify.spec.ts'),
     },
     output: {
       path: __dirname + '/test-build',
